Fixes menu positioning inside of a modal
Fixes: #6

The current implementation of the RelativePositionedPopup uses fixed
positioning to place the popup on the screen. I overlooked the fact that
fixed positioning inside a transformed element will be relative to the
transformed element, but the client bounding rects does not reflect
that.

This is a temporary fix to allow users to use the menu inside a modal.
I'll consider rewriting the positioning logic to use absolute
positioning instead. That also has the advantage of not needing to
repositioning the popup while scrolling.