Make sure that we always inject our CSS before SSC and scope it by version
Currently we have problems when multiple versions of react-components are used
in the same application as the styling can potentially override styles from
self-service-components. This commit makes sure to always inject the style tag
before the self-service-components style tag (update is needed in SSC) and all
CSS is prefixed with the package version.