Add the library version to the CSS hashes
This ensures that multiple versions of react-components does not interfere with
each other.

Example:

<style id='rc-styles'>
<!-- styles for react-components v2-->
<!-- styles for react-components v1-->
</style>

Before when we just content hashed the CSS you could get into situations where a
hash for a class would be the same across versions, that would mean the cascade
would be messed up. The problem I have seen is that the menu `inner` classes is
that same across the two libraries but the menu class is different. When you use
the menu from v2 you accidentally get the `inner` class from v1, that is at the
wrong position in the cascade.