fix(theming): Remove StyleSheetManager from ThemeProvider (#45)
BREAKING CHANGE: The `target` prop on `ThemeProvider` has been removed. In order to do do the same thing that the `target` prop offered you'll need to do the following change:

```
// Old use of `target` prop
<ThemeProvider target={document.body} />
```

The new way is to import the `StyleSheetManager` directly from `styled-components` and wrap that around the `ThemeProvider`

```
// New way to change the location of where styled component css is injected
import { StyleSheetManager } from 'styled-components';

<StyleSheetManager target={document.body}>	
  <ThemeProvider />
</StyleSheetManager>
```

This change will also help with SSR and specifically use in next.js.

Fixes #24.